requirements.txt
# miniQMT量化交易依赖包
# 注意：xtquant需要从QMT安装目录复制到Python环境中

pandas>=1.3.0
numpy>=1.20.0
mysql-connector-python>=8.0.0
PyMySQL>=1.0.0
scipy>=1.7.0
ta-lib>=0.4.0  # 技术分析库
loguru>=0.6.0  # 日志库
python-dotenv>=0.19.0  # 环境变量管理
schedule>=1.1.0  # 任务调度
requests>=2.25.0